+++ Tick-It-Project-Service/src/main/java/com/harish/TickIt/ProjectService/util/JwtUtil.java
@@ -34,10 +34,10 @@ public class JwtUtil
 							   .toList();
 		return roles;
 	}
-	public Integer getEmployeeId(String token)
+	public Long getEmployeeId(String token)
 	{
 		SecretKey secretKey = Keys.hmacShaKeyFor(SECRET_KEY.getBytes());
-		return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("employeeId", Integer.class);
+		return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("employeeId", Long.class);
 	}
 
 }
